added routes and working medications controller without docs

+++ app/controllers/api/v1/medications_controller.rb
@@ -11,6 +11,17 @@ module Api
 				render status: :created, json: meds.to_json
 			end
 
+      def show
+				
+        medication = Medication.find_by_name(params[:name])
+
+				if not medication.blank?
+					render status: :ok, json: medication.to_json
+				else
+					render status: :ok, json: "Medication not found"
+				end
+			end
+      
 			def create
 
 				med = Medication.new(medication_params)
@@ -22,17 +33,34 @@ module Api
 			    end
 			end
 
+    	def update
+       
+         medication = Medication.find_by_name(params[:name])
+
+				if not medication.blank?
+          if medication.update(medication_params)
+            render status: :ok, json: medication.to_json
+          else
+            render status: :bad_request, json: medication.errors
+          end
+				else
+					render status: :ok, json: "Medication not found"
+				end
+      end
+      
+      
 			def destroy
 
-				med = Medication.find(params[:id])
-
-				if med.destroy
-					render status: :ok, json:"Medication Deleted."
-				else
-					render status: :bad_request, json: med.errors
+        med = Medication.find_by_name(params[:name])
+        if not med.blank?
+          if med.destroy
+            render status: :ok, json:"Medication Deleted."
+          end
+        else
+					render status: :ok, json: "Medication not found"
 				end
 			end
-			
+        	
 			private
 
 			def medication_params
